Bug 24341: Allow for more complex markup in OPAC confirmation dialogs
This patch changes the globally-available confirmModal() function in the
OPAC so that it uses jQuery's html() function for adding information to
the confirmation modal's body. This will allow for more than just plain
text to be added.

This patch also removes some obsolete lines from the JavaScript in
opac-user.tt and renames some variable for clarity.

To test, apply the patch log in to the OPAC as a user who has holds and
article requests.

 - Test the following interactions, both the confirmation and cancel
   options for each:
     - Cancel a hold
     - Suspend all holds
     - Resume all holds
     - Cancel an article request

Signed-off-by: Maryse Simard <maryse.simard@inlibro.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>