Bug 24338: Add 'CASH' to default 'PAYMENT_TYPE' AV
Bug 23321 adds cash registers to the koha accounts system and as part of
that work introduces the 'CASH' payment type as a required option to
allow requirment of a 'CASH_REGISTER' when making cash payments. The
original patchset, however, forgot to add this new authorised value
during installation and upgrade.

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Joy Nelson <joy@bywatersolutions.com>